requirements.txt
aiogram<3.0
aiohttp==3.6.2
aioredis<2.0
python-dotenv~=0.19.2
asyncpg~=0.25.0
loader~=2017.9.11